backend/requirements.txt
Django>=4.2
djangorestframework>=3.15
psycopg2-binary>=2.9
gunicorn>=20.1
django-cors-headers>=4.0
